Changed user actions on the client to user service
Also checked the encrypt/decrypt of user ID client side is working
