+++ client/room-booking/package.json
@@ -23,6 +23,7 @@
     "@angular/platform-browser-dynamic": "~7.1.0",
     "@angular/router": "~7.1.0",
     "core-js": "^2.5.4",
+    "crypto-js": "^3.1.9-1",
     "hammerjs": "^2.0.8",
     "rxjs": "~6.3.3",
     "sweetalert2": "^8.0.7",
