+++ client/room-booking/src/services/data.service.ts
@@ -16,39 +16,39 @@ export class DataService {
     this.User = new UserDataModel();
   }
 
-  /**
-  * Registers a new user
-  */
-  public registerUser() {
-    return this.http
-      .post(this.baseUrl + '/api/register', { data: this.User }, { headers: this.headers })
-      .pipe(map((response: any) => response.json()));
-  }
+  // /**
+  // * Registers a new user
+  // */
+  // public registerUser() {
+  //   return this.http
+  //     .post(this.baseUrl + '/api/register', { data: this.User }, { headers: this.headers })
+  //     .pipe(map((response: any) => response.json()));
+  // }
 
-  /**
-  * Logins a user
-  */
-  public loginUser() {
-    return this.http
-      .post(this.baseUrl + '/api/login', { data: this.User }, { headers: this.headers })
-      .pipe(map((response: any) => response.json()));
-  }
+  // /**
+  // * Logins a user
+  // */
+  // public loginUser() {
+  //   return this.http
+  //     .post(this.baseUrl + '/api/login', { data: this.User }, { headers: this.headers })
+  //     .pipe(map((response: any) => response.json()));
+  // }
 
-  /**
-   * Gets user profile data from the server
-   */
-  public getUserProfile() {
-    const userId = this.getUserId();
-    return this.http
-      .get(this.baseUrl + '/api/user/' + userId, { headers: this.headers })
-      .pipe(map((response: any) => response.json()));
-  }
+  // /**
+  //  * Gets user profile data from the server
+  //  */
+  // public getUserProfile() {
+  //   const userId = this.getUserId();
+  //   return this.http
+  //     .get(this.baseUrl + '/api/user/' + userId, { headers: this.headers })
+  //     .pipe(map((response: any) => response.json()));
+  // }
 
-  /**
-   * Gets the User ID from local storage item
-   */
-  public getUserId() {
-    const userId = localStorage.getItem('id');
-    return userId;
-  }
+  // /**
+  //  * Gets the User ID from local storage item
+  //  */
+  // public getUserId() {
+  //   const userId = localStorage.getItem('id');
+  //   return userId;
+  // }
 }
